requirements.txt
pywin32>=306
psutil>=5.9.0
Pillow>=9.0.0
pynput>=1.7.6
keyboard>=0.13.5